Corrects set_payload not working correctly
`set_payload` can only be called once. The first time you call it,
it automatically encodes the message, but for the second time, since
the charset is already set, it doesn't encode it again. Hence, you
need to "reset" it by deleting the Transfer-Encoding header and then
calling set_payload.