Improve handling of "Date" header field
Under normal circumstances the "Date" header is added by mailmerge - just as it always has been.
If "Date" header is already present (from the template) - a new "Date" header is not added.
If "Date" header is already present but empty - then delete it

Summary:
If the "Date" is to be left out - then just add an empty "Date" header into the template

This presents a (backwards compatible) way to let the mail-system add its own "Date" upon queuing into mail-queue.